Added shield to enemy

+++ scripts/health.gd
@@ -6,6 +6,7 @@ class_name Health
 @export var max_health = 3
 @export var immunity_frame_duration = 0.0
 
+var shields = 0
 var last_hit_time : float
 var _is_dead = false
 var current_health : int
@@ -19,9 +20,12 @@ func _ready() -> void:
 func Damage(amount : int):
 	var current_time = Time.get_ticks_msec() / 1000.0
 	if current_time - last_hit_time > immunity_frame_duration and not _is_dead:
+		last_hit_time = current_time
+		if shields > 0:
+			shields -= 1
+			return
 		current_health -= amount
 		animation_player.play("flicker")
-		last_hit_time = current_time
 		send_signal()
 	
 func Heal(amount : int):
